Simplify most exception handling; avoiding unnecessary wrapping and generic checked throws
Also tidies Object.equals and some minor refactorings across the codebase.

+++ server/src/testFixtures/java/com/thoughtworks/go/domain/valuestreammap/VSMTestHelper.java
@@ -111,7 +111,7 @@ public class VSMTestHelper {
         CaseInsensitiveString pipelineName = node.getId();
         List<Revision> revisions = node.revisions();
         for (Revision revision : revisions) {
-            if (revision instanceof PipelineRevision && revision.getRevisionString().equals(new PipelineIdentifier(pipelineName.toString(), Long.parseLong(counter)).pipelineLocator())) {
+            if (revision instanceof PipelineRevision && revision.getRevisionString().equals(new PipelineIdentifier(pipelineName.toString(), Integer.parseInt(counter)).pipelineLocator())) {
                 return (PipelineRevision) revision;
             }
         }
